fix(ci): sever helm-docs image acquisition from the changelog helpers
prepare-release.mjs's renderHelmDocs() (called via RENDER_HELM_DOCS=1)
made every consumer of this file's changelog helpers -- including
verify-changelog-fresh.mjs, which pr-hygiene.yml's pr-body job runs --
read as an image-acquiring script to assert-image-jobs-authenticate.mjs,
which follows every relative import transitively by design. Move the
image-acquiring code to lib/helm-docs.mjs so prepare-release.mjs's own
consumers are no longer coupled to it, and give prepare-release.yml's
prepare job the real GHCR + Docker Hub logins it was genuinely missing
(consolidating its separate, unauthenticated
`uses: docker://jnorwood/helm-docs:...` step into the same call). A
GHCR login on pr-hygiene.yml's conditional changelog-freshness step
satisfies the scanner's remaining, deliberately conservative
over-approximation.
